Just One Return Statement
Excellent work! Your `response_path` method breaks this down just as I would have, and your code is very canonical.

This pull request just demonstrates one guideline for more readable code that you might be interested in. And that guideline is: only have one return statement in a function/method.

The idea is that when someone else is looking at your code, if there are multiple return statements then it's a bit harder to reason out which lines of code got called or executed in the method/function. Having just one return statement gives a clear starting point (the top of the function/method) and ending point (the return statement).

But it's just a guideline because sometimes having a single return statement will make your code less readable. It's just something to keep in mind and apply as you see fit. (: